chore: WorkflowRun 도메인 세팅
