--- apps/user-service/src/main/java/site/icebang/domain/execution/model/WorkflowRun.java
@@ -1,30 +1,0 @@
-package site.icebang.domain.execution.model;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class WorkflowRun {
-
-    private Long id;
-    private Long workflowId;
-    private String traceId;
-    private String runNumber;
-    private String status;
-    private String triggerType;
-    private LocalDateTime startedAt;
-    private LocalDateTime finishedAt;
-    private Long createdBy;
-    private LocalDateTime createdAt;
-
-}
