+++ apps/user-service/src/main/resources/mybatis/mapper/WorkflowRunMapper.xml
@@ -1,21 +1,27 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="site.icebang.domain.execution.mapper.WorkflowRunMapper">
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-    <insert id="save" useGeneratedKeys="true" keyProperty="id">
-        INSERT INTO workflow_run (workflow_id, trace_id, run_number, status, trigger_type, started_at, finished_at, created_by)
-        VALUES (#{workflowId}, #{traceId}, #{runNumber}, #{status}, #{triggerType}, #{startedAt}, #{finishedAt}, #{createdBy})
+<mapper namespace="site.icebang.domain.workflow.mapper.execution.WorkflowRunMapper">
+
+    <resultMap id="WorkflowRunResultMap" type="site.icebang.domain.workflow.model.execution.WorkflowRun">
+        <id property="id" column="id"/>
+        <result property="workflowId" column="workflow_id"/>
+        <result property="traceId" column="trace_id"/>
+        <result property="status" column="status"/>
+        <result property="startedAt" column="started_at"/>
+        <result property="finishedAt" column="finished_at"/>
+        <result property="createdAt" column="created_at"/>
+    </resultMap>
+
+    <insert id="insert" useGeneratedKeys="true" keyProperty="id">
+        INSERT INTO workflow_run (workflow_id, trace_id, status, started_at, created_at)
+        VALUES (#{workflowId}, #{traceId}, #{status}, #{startedAt}, #{createdAt})
     </insert>
 
     <update id="update">
         UPDATE workflow_run
-        <set>
-            <if test="status != null">status = #{status},</if>
-            <if test="startedAt != null">started_at = #{startedAt},</if>
-            <if test="finishedAt != null">finished_at = #{finishedAt},</if>
-        </set>
+        SET status = #{status},
+            finished_at = #{finishedAt}
         WHERE id = #{id}
     </update>
